requirements.txt
pycryptodome
tabulate
intervaltree
hexdump
